move cuda tests from core to cudaarithm
