+++ modules/core/test/test_main.cpp
@@ -7,14 +7,4 @@
 
 #include "test_precomp.hpp"
 
-#ifndef HAVE_CUDA
-
 CV_TEST_MAIN("cv")
-
-#else
-
-#include "opencv2/ts/cuda_test.hpp"
-
-CV_CUDA_TEST_MAIN("cv")
-
-#endif
