+++ modules/cudaarithm/test/test_gpumat.cpp
@@ -40,7 +40,7 @@
 //
 //M*/
 
-#include "../test_precomp.hpp"
+#include "test_precomp.hpp"
 
 #ifdef HAVE_CUDA
